buildenv: Implement croot to get back to SRC_DIR quickly
UN1CA has been growing in the last months and there are lots of modules,
this commit wants to help returning quickly to the top directory
with a single command.

croot:      Changes directory to the top of the tree, or a subdirectory thereof.

This follows the AOSP implementation of croot in envsetup.sh
More details: https://android.googlesource.com/platform/build/+/refs/tags/android-15.0.0_r1/envsetup.sh#806

Change-Id: Ia87963d34455ea1b94687c659749937fe82ad0f1