buildenv: Implement croot to get back to SRC_DIR quickly
UN1CA has been growing in the last months and there are lots of modules,
this commit wants to help returning quickly to the top directory
with a single command.

croot:      Changes directory to the top of the tree, or a subdirectory thereof.

This follows the AOSP implementation of croot in envsetup.sh
More details: https://github.com/LineageOS/android_build/blob/lineage-21.0/envsetup.sh#L1083