unica: mods: eureka: add support for essi
This is the best way to implement this. They would share the same apks and waste twice the space.
Essi doesn't have ACH versions by default so the original script cannot be used.

So directly replace the original media folder with S24 one on both and also avoid wasting
github action minutes to download and extract 20GB a second time.